requirements.txt
acora==2.0
aff4-snappy==0.5.1
antlr4-python3-runtime==4.7.2
arrow==0.10.0
artifacts==20170909
backcall==0.1.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
decorator==4.4.1
expiringdict==1.1.4
fastchunking==0.0.3
filelock==2.0.6
filemagic==1.6
Flask==1.1.1
future==0.18.2
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
intervaltree==2.1.0
ipaddr==2.2.0
ipython==6.5.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
kafka-python==1.4.7
libqcow-python==20191103
MarkupSafe==1.1.1
oauth2client==3.0.0
paramiko==2.6.0
parsedatetime==2.4
parso==0.5.1
pathlib==1.0.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
portpicker==1.1.1
prompt-toolkit==1.0.18
psutil==5.6.6
ptyprocess==0.6.0
pyaff4==0.27
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyBindGen==0.20.1
pyblake2==1.1.2
pycparser==2.19
pycryptodome==3.4.7
pyelftools==0.24
Pygments==2.4.2
PyNaCl==1.3.0
pyparsing==2.1.5
pysftp==0.2.9
python-dateutil==2.8.1
pytsk3==20170802
pytz==2019.3
PyYAML==3.12
rdflib==4.2.2
readline==6.2.4.1
rekall==1.7.1
rekall-agent==1.7.1
rekall-capstone==3.0.5.post2
rekall-core==1.7.2rc1
rekall-efilter==1.6.0
rekall-lib==1.7.2rc1
rekall-yara==3.6.3.1
requests==2.22.0
rsa==4.0
simplegeneric==0.8.1
simplejson==3.17.0
six==1.13.0
sortedcontainers==1.5.7
SPARQLWrapper==1.8.4
sseclient==0.0.18
stix2==1.2.1
stix2-patterns==1.2.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
